Decompose utils file to folder
