+++ src/utils/index.ts
@@ -1,0 +1,2 @@
+export { santizeTitle, santizeTitleExcess } from './santizeTitle';
+export { addOrReplaceFrontMatter } from './frontmatter';
